+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.4.0beta] - 2016-10-xx
+## [0.4.0beta] - 2016-10-03
 ### Added
 - `bower.json` Optional
 - Set boilerplate version on `gulpfile.js`